rename options store
