+++ src/extensionConsole/App.tsx
@@ -16,10 +16,7 @@
  */
 
 import React, { useEffect } from "react";
-import store, {
-  hashHistory,
-  persistor,
-} from "@/extensionConsole/store/optionsStore";
+import store, { hashHistory, persistor } from "@/extensionConsole/store";
 import { Provider, useDispatch } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
 import { Container } from "react-bootstrap";
