+++ src/extensionConsole/pages/mods/modals/convertToModModal/ConvertToModModal.test.tsx
@@ -28,7 +28,7 @@ import {
   selectModalsContext,
   selectShowShareContext,
 } from "@/extensionConsole/pages/mods/modals/modModalsSelectors";
-import { type RootState } from "@/extensionConsole/store/optionsStore";
+import { type RootState } from "@/extensionConsole/store";
 import { authStateFactory } from "@/testUtils/factories/authFactories";
 import { standaloneModDefinitionFactory } from "@/testUtils/factories/modComponentFactories";
 import { timestampFactory } from "@/testUtils/factories/stringFactories";
