+++ src/extensionConsole/pages/mods/modals/modModalsSelectors.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type RootState } from "@/extensionConsole/store/optionsStore";
+import { type RootState } from "@/extensionConsole/store";
 
 export const selectShowLogsContext = ({ modModals }: RootState) =>
   modModals.showLogsContext;
